requirements.txt
sqlalchemy==1.3.17
apache-iotdb==0.12.2
future
more-itertools==8.3.0